refactored reflex_store.js to reflexes.js, received() to process.js
