+++ javascript/reflex_data.js
@@ -1,10 +1,7 @@
 import Schema from './schema'
 
 import { extractElementAttributes, extractElementDataset } from './attributes'
-import { getReflexRoots } from './reflexes'
-import { uuidv4 } from './utils'
-import { elementToXPath } from './utils'
-
+import { uuidv4, getReflexRoots, elementToXPath } from './utils'
 import { version } from '../package.json'
 
 export default class ReflexData {
