+++ javascript/utils.js
@@ -1,3 +1,5 @@
+import Schema from './schema'
+
 // uuid4 function taken from stackoverflow
 // https://stackoverflow.com/a/2117523/554903
 
@@ -139,6 +141,65 @@ const elementInvalid = element => {
   )
 }
 
+const getReflexElement = (args, element) => {
+  return args[0] && args[0].nodeType === Node.ELEMENT_NODE
+    ? args.shift()
+    : element
+}
+
+const getReflexOptions = args => {
+  const options = {}
+  if (
+    args[0] &&
+    typeof args[0] === 'object' &&
+    Object.keys(args[0]).filter(key =>
+      [
+        'attrs',
+        'selectors',
+        'reflexId',
+        'resolveLate',
+        'serializeForm',
+        'suppressLogging',
+        'includeInnerHTML',
+        'includeTextContent'
+      ].includes(key)
+    ).length
+  ) {
+    const opts = args.shift()
+    // TODO: in v4, all promises resolve during finalize stage
+    // if they specify resolveLate, console.warn to say that the option will be ignored
+    // deprecation warning in 3.5 is not required as it's still required until v4
+    Object.keys(opts).forEach(o => (options[o] = opts[o]))
+  }
+  return options
+}
+
+// compute the DOM element(s) which will be the morph root
+// use the data-reflex-root attribute on the reflex or the controller
+// optional value is a CSS selector(s); comma-separated list
+// order of preference is data-reflex, data-controller, document body (default)
+const getReflexRoots = element => {
+  let list = []
+  while (list.length === 0 && element) {
+    let reflexRoot = element.getAttribute(Schema.reflexRoot)
+    if (reflexRoot) {
+      if (reflexRoot.length === 0 && element.id) reflexRoot = `#${element.id}`
+      const selectors = reflexRoot.split(',').filter(s => s.trim().length)
+      if (Debug.enabled && selectors.length === 0) {
+        console.error(
+          `No value found for ${Schema.reflexRoot}. Add an #id to the element or provide a value for ${Schema.reflexRoot}.`,
+          element
+        )
+      }
+      list = list.concat(selectors.filter(s => document.querySelector(s)))
+    }
+    element = element.parentElement
+      ? element.parentElement.closest(`[${Schema.reflexRoot}]`)
+      : null
+  }
+  return list
+}
+
 export {
   uuidv4,
   serializeForm,
@@ -149,5 +210,8 @@ export {
   elementToXPath,
   XPathToElement,
   XPathToArray,
-  elementInvalid
+  elementInvalid,
+  getReflexElement,
+  getReflexOptions,
+  getReflexRoots
 }
